requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
aiohttp==3.9.1
aiosmtplib==3.0.1
email-validator==2.1.0
beautifulsoup4==4.12.2
dnspython==2.4.2
redis==5.0.1
elasticsearch==8.11.0
apscheduler==3.10.4
python-dotenv==1.0.0
cryptography==41.0.7